+++ Whisky/en.lproj/Localizable.strings
@@ -75,6 +75,7 @@
 "kill.bottles" = "Kill All Bottles";
 "open.logs" = "Open Logs Folder";
 "open.bottle" = "Import Bottle";
+"open.setup" = "Setup";
 
 "showAlertOnFirstLaunch.messageText" = "Would you like to move Whisky to your Applications folder?";
 "showAlertOnFirstLaunch.informativeText" = "In some cases, Whisky wont't be able to function properly from a different folder";
